# change type of lookupD