--- a/Base32768/Base32768.csproj
+++ b/Base32768/Base32768.csproj
@@ -18,6 +18,6 @@
     <PackageProjectUrl>https://github.com/naminodarie/Base32768</PackageProjectUrl>
     <RepositoryUrl>https://github.com/naminodarie/Base32768</RepositoryUrl>
 
-    <Version>1.0.1</Version>
+    <Version>1.0.2</Version>
   </PropertyGroup>
 </Project>